InfraS

+++ class/subtotal.class.php
@@ -58,7 +58,11 @@ class TSubtotal {
 			/**
 			 * @var $object Propal Fournisseur
 			 */
-			else if($object->element=='supplier_proposal') $res = $object->addline($desc, 0,$qty,0,0,0,0,0,'HT',0,0,9,$rang, TSubtotal::$module_number, 0, 0, 0, $label);
+			else if($object->element=='supplier_proposal') {
+                /** @var SupplierProposal $object */
+                $object->special_code = TSubtotal::$module_number;
+                $res = $object->addline($label, 0, $qty, 0, 0, 0, 0, 0, 'HT', 0, 0, 9, $rang, TSubtotal::$module_number, 0, 0, 0, '');
+            }
 
 			/**
 			 * @var $object Commande
@@ -727,6 +731,8 @@ class TSubtotal {
 		        break;
 
 		    case 'supplier_proposal':
+                $object->special_code = self::$module_number;
+                if (empty($desc)) $desc = $label;
 		        $res = $object->updateline($rowid, $pu, $qty, $remise_percent, $txtva, $txlocaltax1, $txlocaltax2, $desc, $price_base_type, $info_bits, $special_code, $fk_parent_line, $skip_update_total, $fk_fournprice, $pa_ht, $label, $type, $array_options,'', $fk_unit);
 		        break;
 
@@ -864,7 +870,7 @@ class TSubtotal {
 	 * @param array $parameters assoc array; keys: 'object' (CommonObject), 'file' (string), 'outputlangs' (Translate)
 	 * @param null  $origin_pdf unused [lines that used it are commented out]
 	 */
-	public static function addRecapPage(&$parameters, &$origin_pdf)
+	public static function addRecapPage(&$parameters, &$origin_pdf, $fromInfraS = 0)	// InfraS change
 	{
 		global $user,$conf,$langs;
 
@@ -1108,8 +1114,8 @@ class TSubtotal {
 		$pdf->Close();
 		$pdf->Output($file,'F');
 
-		$pagecount = self::concat($outputlangs, array($origin_file, $file), $origin_file);
-
+		if (empty($fromInfraS))		$pagecount = self::concat($outputlangs, array($origin_file, $file), $origin_file);	// InfraS change
+		if (!empty($fromInfraS))	return $file;	// InfraS add
 		if (empty($conf->global->SUBTOTAL_KEEP_RECAP_FILE)) unlink($file);
 	}
